pyproject.toml
[project]
name = "blindtest"
version = "0.1.0"
description = "A web-based multiplayer music quiz game using Spotify playlists"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Flask>=3.0.0",
    "Flask-SocketIO>=5.3.5",
    "Flask-CORS>=4.0.0",
    "python-socketio>=5.10.0",
    "python-dotenv>=1.0.0",
    "spotipy>=2.23.0",
    "openai>=1.6.1",
    "eventlet>=0.33.3",
]
